JBM-614 continue refactoring and cleaning
* remove unnecessary public modifier
* cast types
* shorten overlong statements
* correct some returns
* move classes to the right location
* etc.

+++ grails-app/services/net/biomodels/jummp/core/NotificationService.groovy
@@ -85,7 +85,7 @@ class NotificationService implements InitializingBean {
 
     Set<User> getNotificationRecipients(def permissionsMap) {
         def writeAccessList = permissionsMap.findAll { ptc -> ptc.write }
-        def recipients = writeAccessList.collect { User.get(it.id) }
+        def recipients = writeAccessList.collect { User.get(it.id as Long) }
         recipients
     }
